feat: Update dependencies and refactor telemetry components for push based metrics exporting.
