+++ packages/operator/Cargo.toml
@@ -5,15 +5,15 @@ edition = "2021"
 
 
 [dependencies]
-kube = { version = "1.1.0", features = ["derive"] }
-kube-runtime = "1.1.0"
+kube = { version = "2.0.1", features = ["derive"] }
+kube-runtime = "2.0.1"
 serde = { version = "1.0.224", features = ["derive"] }
 serde_json = "1.0.142"
 tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread"] }
 thiserror = "2.0.16"
 tracing = "0.1.41"
 k8s-openapi = { version = "0.26.0", features = ["latest", "schemars"] }
-schemars = "0.8"
+schemars = "1.0.0"
 futures = "0.3.31"
 anyhow = "1.0.99"
 mimalloc = "0.1.48"
@@ -21,12 +21,12 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 opentelemetry_sdk = "0.30.0"
 opentelemetry-otlp = "0.30.0"
 opentelemetry-appender-tracing = "0.30.0"
-prometheus = "0.14.0"
 hyper = { version = "1.6.0", features = ["full"] }
 sysinfo = "0.37.0"
 http-body-util = "0.1.3"
 bytes = "1.10.1"
 hyper-util = "0.1.17"
+opentelemetry = "0.30.0"
 
 
 [target.'cfg(target_arch = "x86_64")']
