+++ packages/operator/src/telemetry/mod.rs
@@ -1,2 +1,3 @@
 pub mod logging;
-pub mod telemetry_server;
+pub mod metrics;
+mod resource;
